version: Use `setuptools_scm` to generate the version (bug 1867133)
`setuptools_scm` correctly generates the `version.py` file when it's
built, but it only creates it inside the container. For testing, the
local disk volume overrides the contents of the container, so the
`version.py` file is not accessible for local testing.

As such, we need to explicitly generate the `version.py` file (with a
new command: `lando generate_version_file`) and run it before we start
the server to ensure that it exists.

(Also added `version.py` to the `.gitignore` file.)