create_environment_repos: skip repo when create fails (bug 1936693)
An error message is displayed when this happens.

This may happen when a repository entry is misconfigured, or if the repo
is otherwise misconfigured.

Skipping rather than failing allows to continue starting the Lando
service, where the entry can be manually fixed (or in cases where the
failing repository is unneeded).