repo: rename push_bookmark to push_target in model
We do this to have a less HG-specific term that can be reused for other
repo types. By removing the little bit of indirection provided by the
previous alias method, we lose the None-ification of empty string (which
Django forces on blank-able Char form fields), so we instead handle this
directly within the worker.