automation-actions: implement `merge-onto` action (Bug 1959823) (#291)
Implement the `merge-onto` action for the automation API, which
takes a `target` and `strategy` parameter for the merge. If
the `strategy` isn't passed, a regular merge is performed and
merge conflicts are considered a failure. If the strategy is
"ours", the merge is performed with the current branches content
being taken. If the "theirs" strategy is passed, the target branches
content is preferred instead, by merging into a temp branch with
the `ours` strategy and moving the destination branch onto the
merge commit.