automation-actions: implement `tag` action (Bug 1959583) (#292)
Implement the `tag` automation action. The action creates
a tag at either the currently checked out commit during
action processing, or at a specific commit if passed.

Update the `run_automation_job` logic to track new tags
and push them separately on Git, since Git does not create
new commits for tags.

We also updated `set_landing_system.py` to work with
the `tag` command. Since the `extra` option is passed
explicitly as `None`, using `.setdefault` does not set
the value to a dict, so we update the logic to ensure it
is set appropriately.