phabricator: support limits when collating (bug 1957891)
- ignore limit when calling call_conduit
- return data based on original limit after collating

For queries that are expected to return greater than 100 results
setting the limit does not make sense as Phabricator limits results
to 100 anyway. So in those cases, we ignore the limit set, and then
discard any excess records. This is done for compatibility at this
time since various methods currently set a limit.