git: only force a `--no-ff` merge when strategy is explicitly passed (Bug 1963699)
Currently we have been forcing fast-foward merges in `GitSCM.merge_onto`.
When possible, we want to allow fast-foward merges to take place.
Update `merge_onto` to only enforce `--no-ff` when using a merge
strategy, since in that case the strategy will be avoided when
a fast-foward is possible, so the strategy will be ignored.