create_environment_repo: add releng staging branches (bug 1964674) (#316)
This will allow RelEng to have a set of dedicated repositories that we
can write to as part of Try tasks to verify our own automation, that
we'll also be able to force push to/reset as needed.

We explicitly do _not_ want to sync these back to hg to avoid any
complications when we force push to them.

I ended up using scm level 1 for these, since they are Try-oriented
repositories.