headless_api: add support for RelBranches (Bug 1964716)
Add support for creating and updating commits on RelBranches
to the headless API. Add a new top-level `relbranch` field to
the `AutomationOperation` schema, which takes a `RelBranchSpecifier`
object. This object contains the name of the relbranch to push
changes to, and an optional `commit_sha` which is used as the
base when creating a new RelBranch.

The values in the `relbranch` field are saved as a new JSON
column in the `AutomationJob`. Before running the job, we
reference this field to determine if we should use RelBranch
specific base commits to apply actions to, and a RelBranch
specific push target ref. Since it is possible for the
branch to be new, and therefore not have an upstream tracking
branch to compare to, we update `scm.describe_local_changes`
with an optional argument which indicates the base commit
to compare against to find local changes.