headless_api: add an action to apply patches as binary (Bug 1972956)
Add a new `add-commit-base64` action which accepts a base64
encoded patch in the format of the target repo to apply patches
to. This action will directly apply patches, using `git am` for
Git and `hg import` for Mercurial. Add an `apply_patch_bytes`
method on the `AbstractSCM` class and implementations for each
concrete SCM.