stack: always display "Request Uplift" button when user is logged in (Bug 1976527) (#416)
At the moment the button only displays when the "Preview Landing" button
is displayed. "Preview Landing" is only displayed when the landing is
valid, which means that if a revision is closed and the user is missing
their Phabricator API key, the button will not display.

Move the "Request Uplift" button to a separate `if` block that only
checks if the user is logged in.