treestatus: port from Flask to Django (Bug 1893312) (#442)
This PR migrates the existing Treestatus app from Flask
to Django. The app is fully functional but is not integrated
with the standard Lando flows. For now, the app continues
calling out to Treestatus in old-Lando to determine the
tree statuses for landing.

Treestatus `GET` API endpoints are ported to Django-Ninja
using the existing Pydantic testing schemas in Lando-API.
The previous `PUT`/`PATCH`/`POST`/`DELETE` API endpoints
are changed to Python methods, but no API endpoints are
provided as these methods can be called directly in the
unified Lando. Future support could be added by defining
a new API endpoint which calls the Python methods.

Various Treestatus enum types are converted to Django's
`TextChoices`, which integrates nicely due to being a
subclass of both `str` and `Enum`, and providing a
`choices` field, which removes the need for `to_choices()`.

The Flask implementation of Treestatus had several helper
functions to convert values from WTForms to be JSON serializable
for submission to Lando-API. These helper methods are removed as
they are not necessary in the unified repo.

The status change update form previously had multiple buttons in
the form definition, these are changed to the `action` choice field
in the form.

The `create_new_tree` method now creates a new `Log` entry for
the new tree to avoid an issue where loading the log view for
the tree would fail as no logs are available.

Tests are updated to replace API calls with their equivalent
Python methods. UI tests from Lando-UI are removed as they
existed to test conversions between form submission in
Lando-UI and API submission to Lando-API.