admin: add search_fields for most pages (bug 1955972)
Most notably, and allow to search Jobs and Revisions by revision_id
(without the 'D' prefix). Also, move users from the filter liste to the
search list.

Generally, the search will allow to filter the visible columns, but
there are a few handy, yet not very discoverable options:
- full-description search for Pushlog Commits,
- full-data search for HeadlessApi Automation Actions, and
- repository name search for Workers.

Some of the searches are on unindexed columns, so they may be a bit
slow, but we can deal with those as we find them.