forms: populate repository choices in `__init__` instead of in class definition (Bug 1985106)
The `repository` field on the `UpliftRequestForm` currently has its
`choices` field defined as the result of the `get_uplift_repositories`
method. However the choices are set at class definition time, meaning
whatever is returned by `get_uplift_repositories` during a Lando deploy
are the choices set for every instantiation of the `UpliftRequestForm`.
If `get_uplift_repositories` doesn't return the proper results (due to
a Phabricator operational issue, network issue, etc) then the form will
have no repositories available as choices. Also, if an uplift repo is
added to Phabricator, it won't be made available until Lando is
re-deployed.

Move population of the `choices` parameter from the class definition
into `__init__` so the choices are determined when the form is
instantiated.