repo: add support for (git)attributes override (bug 1984942) (#488)
This adds an `attributes_override` field to the Repo model, defaulting
to `* !diff`. The override is passed to the SCM's `clean_repo` via
`update_repo`. It is only meaningful for Git, in which case the content
of that field will be set as `.git/info/attributes`.

This allows to overrides in-repo `.gitattributes` that may force binary
files to render as text `diff`s rather than `binary` (which is handy for
human consumption, but error-prone when dealing with text encoding, if
the binary part contains non-UTF8 bytes).