uplift: implement uplift request assessment form in Lando (Bug 1982619)
This commit implements the uplift request assessment form in Lando.

First, we implement a model to store the contents of the assessment.
We also implement an `UpliftRevision` model to associate assessment
responses with revisions. For now each assessment is associated with
a single revision, but this will be used more heavily in a follow-up
PR.

We implement the assessment as a Django form, and update the existing
`UpliftRequestForm` for creating new uplift revisions to inherit from
it. We also implement an `UpliftAssessmentEditForm` which is used
to edit the existing form contents.

On creating a new uplift revision, the contents of the form are stored
in the database and associated with the revision. We then take the form
contents and convert to a string appropriate for consumption from the
Conduit API and pass the form as a transaction to the tip revision in
the newly created stack to populate the form in Phabricator.

We add a new "Uplifts" section to the landing page that is rendered
when the revision's repository has the `approval_required` flag set
on it. We display an "Edit Assessment" button in this section that
opens a modal to edit responses to the assessment without creating
a new uplift request, sending the results to Phabricator via a Celery
task on completion. In the future this section will be expanded to
include other uplift related details.

To facilitate testing, a dummy Phabricator API key is added to the
account on the `user` test fixture. We also add the ability to pass
a Phabricator API key to the `PhabricatorDouble.user` method, so
internal calls to the `whoami` endpoint return expected values.