PatchHelpers: add support for binary patches (bug 1983746) (#486)
This change refactors the interface of the PatchHelpers to allow
creating them from either `StringIO` or `BytesIO`.

First, it removes the `__init__` from the interface, and replace it with
two factory class methods: `from_string_io` and `from_bytes_io`. This
then allows the `GitPatchHelper` to not have to handle a `StringIO`
internally, so it can just use un-decoded bytes all the way through.
Bytes are decoded as needed at the `PatchHelper`'s interface boundary
(e.g., `get_diff`).

There is also a new `get_diff_bytes` on the `GitPatchHelper`, which may
get promoted to the interface in the future, when `from_bytes_io` is
implemented for the `HgPatchHelper`.

Also, add some more tests, update all callsites building `PatchHelpers`,
and upgrade type-hints in the helpers to Python>=3.10.

---

* test_scm_helpers: move from test_hgexports
* helpers: upgrade deprecating Python syntax
* helpers: improve PatchHelper interface
* test_scm_helper: add test for binary Git patches (bug 1983746)
* PatchHelper: make is_diff_line public
* PatchHelper: add support for creating PatchHelpers from BytesIO (bug 1983746)
* test_landings: add test for git binary patch