CommitMap: use correct fromchange argument to json-pushes (bug 1988682) (#531)
We were using `fromchangeset`, but the `json-pushes` command expects
`fromchange`.

Aditionally, when catching up, some older commits in m-u may not have a
`git_commit` recorded, or may not even be present in mf/f. We skip over
those with a warning.

Finally, warn if no commits are known, as there are multiple cases where
a `CommitMap.DoesNotExist` can be raised. This allows to differentiate
them at runtime, and offers a way to fix.