uplifts: warn that "Request Uplift" creates new revisions (Bug 1990937)
When using the "Request Uplift" button from an existing uplift revision
in Lando, add a warning indicating that the form will generate new
Phabricator revisions. This helps avoid confusion when a developer
means to fill out the assessment form, but clicks the "Request Uplift"
button instead.

Also add "Edit" to the button text so it is clearer that the button
is used to edit the form.