uplift: implement uplift requests as background jobs (Bug 1988259)
Implement uplift requests as background jobs by turning
the "Request Uplift" button into a submission of work to Lando.
The uplift request form's repository selection is changed from
a single selection to a multi-select of all repos to uplift
the stack to. On submit, a `MultiTrainUpliftRequest` is created
to track the submission, and an `UpliftJob` is created for each
repository. The status of each job is presented in a new uplift
request UI widget that displays on every relevant revision.

`UpliftJobs` apply the uplifted revisions to the target train.
If the patch applies cleanly, `moz-phab uplift` will be called
to generate new Phabricator revisions using the user's API key.
In the event of a merge conflict the job is marked as a failure
and conflicts are stored on the job for display. Notification
emails are sent on job completion with instructions on next steps.

After this change much of the code around uplift creation is no
longer used and is removed. We also remove the limit of 5 uplift
revisions per request, as that limit was due to the risk of
a connection failure resulting in half-created uplift stacks.

This PR also resolves bug 1850408 by implementing tests for
uplift creation endpoint in the new workflow.