assessments: update existing row instead of create/delete (Bug 1996394) (#639)
Currently when updating existing uplift assessments, we create
a new assessment form from the `request.POST` data without
pre-populating an existing instance in the form. This results
in the `.save()` call creating a new form instance, and we later
delete the old assessment form. This causes issues in the uplift
jobs flow as the form is associatd with multiple revisions, and
deleting the old form woud remove the form from any other revisions
created as part of an uplift request.

To work around this, switch the uplift assessment updating endpoint
to take the revision ID in the URL path, so we can look for an
existing `UpliftRevision` association before validating the form.
This allows us to remove the code to re-associate revisions with
the new uplift form.

Since `revision_id` is now passed in the URL, we can remove the form
field, which allows us to completely remove the
`UpliftEditAssessmentForm`
and use the `UpliftAssessmentForm` directly.