uplifts: provide a way to re-use existing uplift assessment responses (Bug 1998844) (#666)
Adds a new "Link Existing Assessment" button next to the "Edit Uplift
Assessment" button for uplift revisions. This button presents a form
with a single `ModelChoiceField` that references a previously submitted
`UpliftAssessment`. The user can select one of their previous responses
and associate it with the current revision. This allows users to link
a stack of uplift revisions against a new train to the uplift assessment
form responses from uplifting that same stack to a different train.