formatting: use `djlint` for template formatting (Bug 1999751)
Add `djlint` as a `code-quality` requirement to the project.
Run `djlint . --reformat` on the repo root, with the `jinja`
profile and `line_break_after_multiline_tag` enabled. Then
add a `test_djlint` to `test_formatting.py`, to enforce
a passing run of `djlint --check`. Add a `djlint` run to
`lando format`/`make format` so `djlint` is run as a regular
formatting task.