api/views: add PullRequestChecksAPIView, and use landing checks (bug 1992863) (#627)
This change introduces an API endpoint at
https://lando.test/api/pulls/<REPO>/<PR_NO>/checks which provides
current blockers and warnings for the target PR.

Currently, only blockers based on diffs from commits on the PR are
checked.

For test purposes, it makes wider use of the `MockResponse` object,
making it available as a `mock_response` fixture. To fix an interface
mismatch with the `requests.Response` object, the `json` property is
now a method. Tests have been updated accordingly.

---------

Co-authored-by: Zeid <2043828+zzzeid@users.noreply.github.com>
Co-authored-by: Zeid <zeid@mozilla.com>

---------

 * tests: add global mock_response fixture
 * tests: fix MockResponse.json()
 * github: add PullRequestPatchHelper
 * api/views: add PullRequestChecksAPIView (bug 1992863)
 * PullRequestChecksAPIView: reuse LandingChecks for blockers (bug 1992863)
 * views: use pull_request.author to get email for LandingCheck
 * landing_checks: use chk.name() everywhere
 * github.PullRequestPatchHelper: use PullRequest.author to get author details
 * test_github: add commits support to _github_api_get fixture
 * landing_checks: move CONSTs to their respective classes