pull_requests: return 404 on invalid repo or PR (bug 2002104) r=zeid
This PR updates the base GitHub utility to raise any HTTP error as exceptions, so they bubble up to higher layers.

It then captures those in the Pull Request UI view to detect non-existent PRs, and return a cleaner 404 with appropriate message. It also does the same for non-existent Repos.

The Jinja template is also updated to include the exception message, if present, to present more context to the user.

Pull request: https://github.com/mozilla-conduit/lando/pull/741