uplifts: provide a URL to link uplift revisions to assessment forms (Bug 2006101)
Add a new URL for submitting and linking assessment forms to multiple
revisions at once. This will enable `moz-phab` to open the browser
to a page where the uplift request form can be submitted a single
time for multiple revisions concurrently.

When submitting an uplift assessment for two trains:
    /uplift/request/?revisions=1234,1235

When uplifting to a new train:
    /uplift/request/?revisions=1234,1235&assessment_id=5